Use middleware-based filename parser (so we can use functions)
Also adds experimental support for custom parser behavior via a kfconfig.js file. If found, the file applies to media in the same folder and any subfolders. If a subfolder also has a kfconfig.js file it takes priority over the parent's. The file can export (using module.exports) an object or function.

If an object is provided it is used as configuration for the default middleware chain. Currently there are two configurable properties, "separator" (string) and "artistOnLeft" (bool).

If a function is provided it should return an array of middleware functions to run instead of the default chain. The function is passed the default chain as an object so you can still use it while   insertting or appending your own parsing functions. The passed object has three properties: "preMiddleware", "parseMiddleware", and "postMiddleware" that are each arrays containing the default functions for those stages.