QueueItems are now star-able and swipe-able
If the item has a remove button, the star button is behind a swipe. I think a user is more likely to star a song after it's been played, and multiple buttons visible at once by default looks very busy. In the future an info button should be added (also behind the swipe)