Revert "feat: Support collapsed chrome in Mobile Safari (iOS 15)"
This broke in some release after 15.0. As of 15.4 the bottom browser chrome hides for a second, then pops back. It's actually pretty janky and makes one wonder if they quickly hacked it to fix some popular site that was breaking under the original behavior. Anyway, it sucks and needs more investigation.