fix: use-resize-observer might return undefined
On a view where the Header component doesn't render anything (e.g. Queue or Account when not admin) its height might be undefined, which in the Library view eventually gets passed to react-window, where that is not handled well in its getItemMetadata(), eventually causing an infinite loop/crash in its findNearestItemBinarySearch().