Add CLI KES_PLAYER_DISABLE_WEBGL
The purpose here is to disable visualizations and restore the CDG's original background. This is important for some CDGs that have a complex background or that have multiple singers lyrics in different colors that happen to be the same as the most recent background command.

When KES_PLAYER_DISABLE_WEBGL is not defined or "0" nothing changes.
When KES_PLAYER_DISABLE_WEBGL is any number other than "0", WEBGL is disabled.